remove few serde_json, error handling and cleanup
