--- quadratic-client/src/app/helpers/electronMaximizeCurrentWindow.ts
@@ -1,4 +1,0 @@
-export const electronMaximizeCurrentWindow = () => {
-  //@ts-expect-error
-  if (window.electronAPI) window.electronAPI.maximizeCurrentWindow();
-};
