+++ quadratic-client/src/app/ui/loading/QuadraticLoading.tsx
@@ -1,7 +1,6 @@
 import { ThemeAppearanceModeEffects } from '@/shared/hooks/useThemeAppearanceMode';
 import { Box, LinearProgress } from '@mui/material';
 import { useEffect, useState } from 'react';
-import { TopBarLoading } from '../components/TopBarLoading';
 import './styles.css';
 
 export function QuadraticLoading() {
@@ -26,7 +25,6 @@ export function QuadraticLoading() {
     <>
       {/* ToBarLoading allows window to be moved while loading in electron */}
       <ThemeAppearanceModeEffects />
-      <TopBarLoading></TopBarLoading>
       <div
         style={{
           height: '100%',
