--- quadratic-client/src/shared/utils/isElectron.ts
@@ -1,28 +1,0 @@
-// https://github.com/electron/electron/issues/2288
-export const isElectron = () => {
-  // Renderer process
-  if (
-    typeof window !== 'undefined' &&
-    typeof window.process === 'object' &&
-    //@ts-expect-error
-    window.process.type === 'renderer'
-  ) {
-    return true;
-  }
-
-  // Main process
-  if (typeof process !== 'undefined' && typeof process.versions === 'object' && !!process.versions.electron) {
-    return true;
-  }
-
-  // Detect the user agent when the `nodeIntegration` option is set to false
-  if (
-    typeof navigator === 'object' &&
-    typeof navigator.userAgent === 'string' &&
-    navigator.userAgent.indexOf('Electron') >= 0
-  ) {
-    return true;
-  }
-
-  return false;
-};
